MEN-5259: fix backend tests with the latest deployments-enterprise
Changelog: none
